LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-length-percent-serialize-expected.txt


PASS e.style['margin-left'] = "min(1px + 1%)" should set the property value 
FAIL e.style['margin-left'] = "min(1cm + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1cm)" but got "min(1% + 37.79527559055118px)"
FAIL e.style['margin-left'] = "min(1mm + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1mm)" but got "min(1% + 3.7795275590551185px)"
FAIL e.style['margin-left'] = "min(1Q + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1q)" but got "min(1% + 0.9448818897637796px)"
FAIL e.style['margin-left'] = "min(1in + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1in)" but got "min(1% + 96px)"
FAIL e.style['margin-left'] = "min(1pc + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1pc)" but got "min(1% + 16px)"
FAIL e.style['margin-left'] = "min(1pt + 1%)" should set the property value assert_equals: serialization should be canonical expected "min(1% + 1pt)" but got "min(1% + 1.3333333333333333px)"
PASS e.style['margin-left'] = "min(1em + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1ex + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1ch + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1rem + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1vh + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1vw + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1vmin + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(1vmax + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1px + 1%)" should set the property value 
FAIL e.style['margin-left'] = "max(1cm + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1cm)" but got "max(1% + 37.79527559055118px)"
FAIL e.style['margin-left'] = "max(1mm + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1mm)" but got "max(1% + 3.7795275590551185px)"
FAIL e.style['margin-left'] = "max(1Q + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1q)" but got "max(1% + 0.9448818897637796px)"
FAIL e.style['margin-left'] = "max(1in + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1in)" but got "max(1% + 96px)"
FAIL e.style['margin-left'] = "max(1pc + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1pc)" but got "max(1% + 16px)"
FAIL e.style['margin-left'] = "max(1pt + 1%)" should set the property value assert_equals: serialization should be canonical expected "max(1% + 1pt)" but got "max(1% + 1.3333333333333333px)"
PASS e.style['margin-left'] = "max(1em + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1ex + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1ch + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1rem + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1vh + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1vw + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1vmin + 1%)" should set the property value 
PASS e.style['margin-left'] = "max(1vmax + 1%)" should set the property value 
PASS e.style['margin-left'] = "min(20px, 10%)" should set the property value 
PASS e.style['margin-left'] = "min(1em, 10%)" should set the property value 
PASS e.style['margin-left'] = "max(20px, 10%)" should set the property value 
PASS e.style['margin-left'] = "max(1em, 10%)" should set the property value 
PASS e.style['margin-left'] = "min(10%, 20px)" should set the property value 
PASS e.style['margin-left'] = "min(10%, 1em)" should set the property value 
PASS e.style['margin-left'] = "max(10%, 20px)" should set the property value 
PASS e.style['margin-left'] = "max(10%, 1em)" should set the property value 
PASS e.style['margin-left'] = "min(10% + 30px, 5% + 60px)" should set the property value 
PASS e.style['margin-left'] = "max(10% + 2em, 5% + 1em)" should set the property value 
FAIL e.style['margin-left'] = "calc(min(10%) + max(1em) + min(20px))" should set the property value assert_equals: serialization should be canonical expected "calc(min(10%) + max(1em) + min(20px))" but got "calc(10% + 1em + 20px)"
FAIL e.style['margin-left'] = "calc(max(20px) + min(1em) + max(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(max(20px) + min(1em) + max(10%))" but got "calc(10% + 1em + 20px)"
FAIL e.style['margin-left'] = "calc(max(10%) + min(1em) + max(20px))" should set the property value assert_equals: serialization should be canonical expected "calc(max(10%) + min(1em) + max(20px))" but got "calc(10% + 1em + 20px)"
FAIL e.style['margin-left'] = "calc(min(20px) + max(1em) + min(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(min(20px) + max(1em) + min(10%))" but got "calc(10% + 1em + 20px)"
FAIL e.style['margin-left'] = "calc(20px + min(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(20px + min(10%))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(10% + min(20px))" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(20px))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(1em + min(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(1em + min(10%))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(10% + min(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(1em))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(min(10%) + 20px)" should set the property value assert_equals: serialization should be canonical expected "calc(20px + min(10%))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(min(20px) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(20px))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(min(10%) + 1em)" should set the property value assert_equals: serialization should be canonical expected "calc(1em + min(10%))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(min(1em) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + min(1em))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(20px + max(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(20px + max(10%))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(10% + max(20px))" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(20px))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(1em + max(10%))" should set the property value assert_equals: serialization should be canonical expected "calc(1em + max(10%))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(10% + max(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(1em))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(max(10%) + 20px)" should set the property value assert_equals: serialization should be canonical expected "calc(20px + max(10%))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(max(20px) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(20px))" but got "calc(10% + 20px)"
FAIL e.style['margin-left'] = "calc(max(10%) + 1em)" should set the property value assert_equals: serialization should be canonical expected "calc(1em + max(10%))" but got "calc(10% + 1em)"
FAIL e.style['margin-left'] = "calc(max(1em) + 10%)" should set the property value assert_equals: serialization should be canonical expected "calc(10% + max(1em))" but got "calc(10% + 1em)"
